vector/requirements.txt
scikit-learn>=1.0.0
numpy>=1.21.0
jieba>=0.42.1
faiss-cpu>=1.7.0 